requirements-gpu.txt
absl-py==0.7.0
adal==1.2.1
ansiwrap==0.8.4
asn1crypto==0.24.0
astor==0.7.1
autograd==1.2
azure-common==1.1.18
azure-datalake-store==0.0.41
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
backcall==0.1.0
bleach==3.1.0
boto3==1.9.88
botocore==1.12.88
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
checksumdir==1.1.5
Click==7.0
cloudpickle==0.7.0
cryptography==2.5
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
defusedxml==0.5.0
docutils==0.14
entrypoints==0.3
future==0.17.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
joblib==0.13.1
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
metakernel==0.20.14
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
numpy==1.16.1
oct2py==4.0.6
octave-kernel==0.28.4
opencv-python==4.0.0.21
pandas==0.24.1
pandocfilters==1.4.2
papermill==0.17.1
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
requests==2.21.0
s3transfer==0.1.13
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
testpath==0.4.2
textwrap3==0.9.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.30.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
